Replace sjson w/ lift-json, change resource format
 * The messagesSent property now contains the original List[int] data
 * SystemInformation uses the same format as WorkerStatistics

+++ src/main/scala/com/signalcollect/interfaces/MessageBus.scala
@@ -36,7 +36,7 @@ trait MessageBus[@specialized(Int, Long) Id, @specialized(Int, Long, Float, Doub
   
   def numberOfWorkers: Int
 
-  def messagesSent: Array[Int]
+  def messagesSent: Array[Long]
   def messagesReceived: Long
 
   def getReceivedMessagesCounter: AtomicInteger
